% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_analogue_distances.R
\name{analogue_distances}
\alias{analogue_distances}
\alias{analogue_distances}
\alias{analogue_distances}
\alias{autoplot.analogue_distances}
\title{ggplot-based plot for analogue distance}
\usage{
analogue_distances(spp, fos, df, x_axis, quantiles = c(0.05, 0.1), ...)

\method{autoplot}{analogue_distances}(object, fill = c("salmon",
  "lightyellow", "skyblue"), categories = c("Good", "Fair", "Poor"), ...)
}
\arguments{
\item{spp}{data.frame or matrix of calibration set species abundances}

\item{fos}{data.frame or matrix of fossil species abundances}

\item{df}{data.frame containing e.g. ages or depths of samples}

\item{x_axis}{character; name of variable for x-axis. Defaults to sample number.}

\item{quantiles}{numeric; quantiles of the training set residual lengths to plot}

\item{...}{extra arguments to \code{\link[rioja]{MAT}}. \code{dist.method} is the only relevant one.}

\item{object}{analogue_distance object}

\item{fill}{character; colours for the background fill for each quality categories}

\item{categories}{character; names of residuals length quality categories.}
}
\value{
Returns a ggplot object.
}
\description{
Calculate and plot the distance to nearest analogue.


...
}
\examples{

data(ImbrieKipp, V12.122, package = "analogue")
## squared residual lengths for Core V12.122
AD <- analogue_distances(ImbrieKipp, V12.122)
autoplot(AD)

}
\author{
Richard J. Telford
}
