% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.residLen.R
\name{autoplot.residLen}
\alias{autoplot.residLen}
\alias{fortify.residLen}
\alias{autoplot.residLen}
\title{ggplot-based plot for objects of class \code{'reslen'}}
\usage{
\method{fortify}{residLen}(object, df)

\method{autoplot}{residLen}(object, df, x_axis, quantiles = c(0.9, 0.95),
  fill = c("salmon", "lightyellow", "skyblue"), categories = c("Good",
  "Fair", "Poor"))
}
\arguments{
\item{object}{an object of class \code{"residLen"}, the result of a call to \code{\link[analogue]{residLen}}}

\item{df}{data.frame containing e.g. ages or depths of samples}

\item{x_axis}{character; name of variable for x-axis. Defaults to sample number.}

\item{quantiles}{numeric; quantiles of the training set residual lengths to plot}

\item{fill}{character; colours for the background fill for each quality categories}

\item{categories}{character; names of residuals length quality categories.}
}
\value{
Returns a ggplot object.
}
\description{
Produces a ggplot object representing the output of objects produced by \code{\link[analogue]{residLen}}.
}
\examples{

data(ImbrieKipp, SumSST, V12.122, package = "analogue")
## squared residual lengths for Core V12.122
rlens <- residLen(ImbrieKipp, SumSST, V12.122)
autoplot(rlens, df = data_frame(age = as.numeric(rownames(V12.122))), x_axis = "age") +
labs(x = "Age", y = "Squared residual distance", fill = "Goodness of fit")
}
\author{
Richard J. Telford
}
