% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_analogue_distances.R
\name{plot_analogue_distances}
\alias{plot_analogue_distances}
\title{ggplot-based plot for analogue distance}
\usage{
plot_analogue_distances(spp, fos, df, x_axis, quantiles = c(0.05, 0.1),
  fill = c("salmon", "lightyellow", "skyblue"), categories = c("Good",
  "Fair", "Poor"), ...)
}
\arguments{
\item{spp}{data.frame or matrix of calibration set species abundances}

\item{fos}{data.frame or matrix of fossil species abundances}

\item{df}{data.frame containing e.g. ages or depths of samples}

\item{x_axis}{character; name of variable for x-axis. Defaults to sample number.}

\item{quantiles}{numeric; quantiles of the training set residual lengths to plot}

\item{fill}{character; colours for the background fill for each quality categories}

\item{categories}{character; names of residuals length quality categories.}

\item{...}{extra arguments to \code{\link[rioja]{MAT}}. \code{dist.method} is the only relevant one.}
}
\value{
Returns a ggplot object.
}
\description{
Produces a ggplot object representing the output of objects produced by \code{\link[analogue]{residLen}}.


...
}
\examples{

data(ImbrieKipp, V12.122, package = "analogue")
## squared residual lengths for Core V12.122
plot_analogue_distances(ImbrieKipp, V12.122)

}
\author{
Richard J. Telford
}
