% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chron_info.R
\name{plot_chron_info}
\alias{plot_chron_info}
\title{Secondary scale for strat.plot}
\usage{
plot_chron_info(x, yvar, dates, xLeft = 0, cex = 0.8)
}
\arguments{
\item{x}{a stratigraphic diagram object produced by strat.plot.}

\item{yvar}{y variable. Same as main plot, probably depth}

\item{dates}{data.frame with columns `up``, `down`, `age`, `error`}

\item{xLeft}{Position of plot. See \code{\link[rioja]{strat.plot}}}

\item{cex}{font size}
}
\description{
Adds a secondary scale for \code{\link[rioja]{strat.plot}}
}
\details{
secondary_scale
}
\examples{
library(rioja)
library(tibble)
data(RLGH)
fos <- RLGH$spec
chron <- RLGH$depths
dates <- tribble(~up, ~down, ~age, ~error,
                  4,      5,   1970,   5,
                  15,     20,  1900,  20)
pt <- strat.plot(fos, yvar = chron$Depth, y.rev = TRUE, xLeft = 0.3, scale.percent = TRUE)
plot_chron_info(pt, yvar = chron$Depth, dates = dates)
}
