% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_scale.R
\name{secondary_scale}
\alias{secondary_scale}
\title{Secondary scale for strat.plot}
\usage{
secondary_scale(
  x,
  yvar,
  yvar2,
  n = 5,
  xLeft = 0.11,
  ylabel2 = "",
  cex.ylabel2 = 1,
  cex.yaxis = 0.8
)
}
\arguments{
\item{x}{a stratigraphic diagram object produced by strat.plot.}

\item{yvar}{y variable. Same as main plot.}

\item{yvar2}{Secondary y variable}

\item{n}{Suggested number of labels on secondary y axis. Defaults to five.}

\item{xLeft}{Position of plot. See \code{\link[rioja]{strat.plot}}}

\item{ylabel2}{character. Label for secondary scale}

\item{cex.ylabel2}{numeric. Size of axis label.}

\item{cex.yaxis}{numeric. Size of axis values.}
}
\description{
Adds a secondary scale for \code{\link[rioja]{strat.plot}}
}
\details{
secondary_scale
}
\examples{
library(rioja)
data(RLGH)
fos <- RLGH$spec
chron <- RLGH$depths
pt <- strat.plot(fos, yvar = chron$Depth, y.rev = TRUE, xLeft = .22, scale.percent = TRUE)
secondary_scale(pt, yvar = chron$Depth, yvar2 = chron$Age, n = 5, ylabel2 = "Years")
}
