% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plot.R
\name{summary_plot}
\alias{summary_plot}
\title{Secondary scale for strat.plot}
\usage{
summary_plot(
  x,
  g,
  yvar,
  xLeft = 0.8,
  xRight = 1,
  y.axis = FALSE,
  ylabel = "",
  cex.ylabel = 1,
  x.name.pos,
  srt.xlabel = 90,
  cex.xlabel = 1,
  fill
)
}
\arguments{
\item{x}{a stratigraphic diagram object produced by strat.plot.}

\item{g}{matrix or data.frame with summary groups. Column names will be used as labels.}

\item{yvar}{y variable. Same as main plot.}

\item{xLeft}{Position of plot. See \code{\link[rioja]{strat.plot}}}

\item{xRight}{ditto}

\item{y.axis}{ditto}

\item{ylabel}{ditto}

\item{cex.ylabel}{ditto}

\item{x.name.pos}{numeric. Position of names. Defaults to midpoint of uppermost sample.}

\item{srt.xlabel}{See \code{\link[rioja]{strat.plot}}}

\item{cex.xlabel}{ditto}

\item{fill}{vector of colours. Defaults to RcolourBrewer Set1}
}
\description{
Adds a secondary scale for \code{\link[rioja]{strat.plot}}
}
\details{
secondary.scale
}
\examples{
 library(rioja)
 library(dplyr)
 data(RLGH)
 fos <- RLGH$spec
 chron <- RLGH$depths

 groups <- case_when(
   substr(names(fos), 1, 1) \%in\% c("A", "E") ~ "Vowel",
   substr(names(fos), 1, 1) \%in\% c("B", "C", "F") ~ "BCF",
   TRUE ~ "NPT"
 ) \%>\%
   factor(levels = c("Vowel", "BCF", "NPT"))

  g <- sapply(levels(groups), function(lev){
    rowSums(fos[, groups == lev, drop = FALSE])
    })


 pt <- strat.plot(fos, yvar = chron$Depth, y.rev = TRUE,
                  xLeft = .1, xRight = 0.8, scale.percent = TRUE)

 summary_plot(pt, g = g, yvar = chron$Depth, xLeft = 0.8)
}
