% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicted_measured_plots.R
\name{fortify.WA}
\alias{fortify.WA}
\alias{fortify.WAPLS}
\alias{fortify.MLRC}
\alias{fortify.MAT}
\alias{fortify.tf}
\alias{autoplot.WAPLS}
\alias{autoplot.WA}
\alias{autoplot.MAT}
\alias{autoplot.MLRC}
\alias{autoplot.tf}
\title{autoplot for rioja transfer functions}
\usage{
\method{fortify}{WA}(object, column = "WA.inv", ...)

\method{fortify}{WAPLS}(object, npls = 1, ...)

\method{fortify}{MLRC}(object, ...)

\method{fortify}{MAT}(object, k = 5, weighted = FALSE...)

\method{fortify}{tf}(object, column, ...)

\method{autoplot}{WAPLS}(object, npls, ...)

\method{autoplot}{WA}(object, ...)

\method{autoplot}{MAT}(object, k = 5, weighted = FALSE, ...)

\method{autoplot}{MLRC}(object, ...)

\method{autoplot}{tf}(object, show_apparent = FALSE, residuals = FALSE,
  smooth = TRUE, ...)
}
\arguments{
\item{object}{transfer function object from rioja}

\item{column}{character, name of column to plot}

\item{npls}{integer, WAPLS component to plot}

\item{k}{integer, MAT number of analogues to plot}

\item{weighted}{logical, plot weighted MAT results?}

\item{show_apparent}{logical, show apparent performance as well as cross-validated}

\item{residuals}{logical, show cross-validated residuals}

\item{smooth}{logical, add smooth to plot}
}
