% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_scale.R
\name{secondary_scale}
\alias{secondary_scale}
\title{Secondary scale for strat.plot}
\usage{
secondary_scale(yvar, yvar2, ylim = NULL, n = 5, y.rev = FALSE,
  xLeft = 0.11, yBottom = 0.07, yTop = 0.8, ylabel2 = "",
  cex.ylabel2 = 1)
}
\arguments{
\item{yvar}{y variable. Same as main plot.}

\item{yvar2}{Secondary y variable}

\item{ylim}{y limits. Defaults to range of data.}

\item{n}{Suggested number of labels on secondary y axis. Defaults to five.}

\item{y.rev}{logical. Reverse y-axis. Do as in main plot.}

\item{xLeft}{Position of plot. See \code{\link[rioja]{strat.plot}}}

\item{yBottom}{ditto}

\item{yTop}{ditto}

\item{ylabel2}{character. Label for secondary scale}

\item{cex.ylabel2}{numeric. Size of axis label.}
}
\description{
Adds a secondary scale for \code{\link[rioja]{strat.plot}}
}
\details{
secondary_scale
}
\examples{
library(rioja)
data(RLGH)
fos <- RLGH$spec
chron <- RLGH$depths
strat.plot(fos, yvar = chron$Depth, y.rev = TRUE, xLeft = .22, scale.percent = TRUE)
secondary_scale(yvar = chron$Depth, yvar2 = chron$Age, n = 5, y.rev = TRUE, ylabel2 = "Years")
}
