% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_to_latin.R
\name{species_to_italics}
\alias{species_to_italics}
\title{Italicise species names}
\usage{
species_to_italics(x, families = TRUE, modifiers = c("spp", "sp", "fo",
  "var", "agg"), postfix = c("-type", "-t"), etc = character(0),
  first_with_dot_only = TRUE, first_capital_only = TRUE,
  no_numbers = TRUE, no_extra_capitals = TRUE, text = TRUE)
}
\arguments{
\item{x}{character; one or more species names to be italicised}

\item{families}{logical; should family names be italicised}

\item{modifiers}{character; vector of taxonomic terms not to italicise}

\item{postfix}{character; text appended to species names that should not be italicised}

\item{etc}{character; vector of other words that should not be italicised}

\item{first_with_dot_only}{logical;}

\item{first_capital_only}{logical; if TRUE only the first word starting with a capital letter is italicised}

\item{no_numbers}{logical; if TRUE any words with numbers are not italicised}

\item{no_extra_capitals}{logical; if TRUE any words with capitals other than first letter are not italicised}

\item{text}{logical; TRUE gives output as character vector; FALSE gives list of expressions}

\item{modifiers_with_dot}{logical, should modifiers end in a dot. Use NA to accept modifiers either with or without a trailing dot.}
}
\description{
Italicises species names, but not modifiers such as "var."
}
\examples{
species_to_italics(c("Navicula spp.", "Navicula spp", "Poaceae"))
library(ggplot2)
data(SWAP, package = "rioja")
SWAP$spec \%>\% mutate(pH = SWAP$pH) \%>\%
pivot_longer(cols = -pH, names_to = "taxa", values_to = "percent") \%>\%
  left_join(as_tibble(SWAP$names), by = c("taxa" = "CODE")) \%>\%
  group_by(taxa) \%>\%
  filter(sum(percent > 0) > 60, max(percent > 20)) \%>\%
  ggplot(aes(x = pH, y = percent)) +
  geom_point() +
  facet_wrap(~TaxonName,
             labeller = as_labeller(
               species_to_italics,
               default = label_parsed))
}
